Renamed trafid_cipher.py to trifid_cipher.py and added doctests
