+++ DIRECTORY.md
@@ -119,7 +119,7 @@
   * [Shuffled Shift Cipher](ciphers/shuffled_shift_cipher.py)
   * [Simple Keyword Cypher](ciphers/simple_keyword_cypher.py)
   * [Simple Substitution Cipher](ciphers/simple_substitution_cipher.py)
-  * [Trafid Cipher](ciphers/trafid_cipher.py)
+  * [Trifid Cipher](ciphers/trifid_cipher.py)
   * [Transposition Cipher](ciphers/transposition_cipher.py)
   * [Transposition Cipher Encrypt Decrypt File](ciphers/transposition_cipher_encrypt_decrypt_file.py)
   * [Vigenere Cipher](ciphers/vigenere_cipher.py)
